fix(AVO-3004): revert previous commit, move hook, add dedicated translations when assignment expires
