+++ src/admin/admin.routes.tsx
@@ -7,8 +7,8 @@ import { renderErrorRoutes } from '../error/error.routes';
 
 import { renderAdminAssignmentRoutes } from './assignments/assignment.routes';
 import { renderCollectionOrBundleRoutes } from './collectionsOrBundles/collections-or-bundles.routes';
-import { renderAdminContentPageLabelRoutes } from './content-page-labels/content-page-label.routes';
 import { renderAdminContentPageRoutes } from './content-page/content-page.routes';
+import { renderAdminContentPageLabelRoutes } from './content-page-labels/content-page-label.routes';
 import { renderAdminDashboardRoutes } from './dashboard/dashboard.routes';
 import { renderInteractiveTourRoutes } from './interactive-tour/interactive-tour.routes';
 import { renderItemRoutes, renderPublishItemRoutes } from './items/items.routes';
