+++ src/shared/components/BlockList/blocks/CollectionFragmentTypeItem.tsx
@@ -1,4 +1,5 @@
 import { DefaultProps, IconName } from '@viaa/avo2-components';
+import { type Avo } from '@viaa/avo2-types';
 import classNames from 'classnames';
 import React, { FC, useMemo } from 'react';
 
@@ -18,7 +19,6 @@ import CollapsibleColumn from '../../CollapsibleColumn/CollapsibleColumn';
 import TextWithTimestamps from '../../TextWithTimestamp/TextWithTimestamps';
 
 import './CollectionFragmentTypeItem.scss';
-import { type Avo } from '@viaa/avo2-types';
 
 export interface CollectionFragmentTypeItemProps extends DefaultProps {
 	block: Avo.Core.BlockItemBase;
